[TreeView] Extract some logic outside of the useTreeView hook
